fix #9
Admin user now counts as a moderator, this extends to the /moderator
page as well as the main page. This privilege gives admin user the
"Moderatoren" link on / as well as a privilege to user
/moderator. Furthermore /moderator now requires moderator
privilege. Previously every logged on user could see the list of
moderators, albeit he had no direct link to that page.